Split up conversion_functions.py
